Animating graph creation
Created a MObject[] (Maybe an ArrayList may function better due to the 0 lines not being added and other complications with floating point arithmetic?). If a Line.display() in the MObject[] returns false, complete will be set to false. This will make sure all Lines are rendered and animated before any other directions are begun.