Added Ellipse class
Since a circle is just a specific case of an ellipse where the a and b coefficients are equal, it is better to repurpose the Circle class as an extension of the Ellipse class. Also added more directions and a sinusoidal interpolation of the tangent line and the circle's size (upcoming).